Carry field values as doubles end to end
A field whose values differ below float's resolution rendered flat: the
plane narrowed each sample to float, so a mu field holding two doubles
1e-17 apart collapsed to one value before any range or colour code ran.
Widen the slice, line, page and volume containers to double and drop the
narrowing; it survives only to encode for a pre-1.5 peer.

Protocol 1.5 pairs each value vector with a values_f64 rather than
widening it in place, which a 1.4 peer would misread as half a vector.
Slice cells cost 11 bytes now, so the default frame budget rises to 192
MiB to keep a 4096^2 raster fitting, and the volume grid cache to 512
MiB to still hold the four grids it documents.

+++ tests/unit/test_session_validation.cpp
@@ -1310,7 +1310,7 @@ int main()
 
     // --- the Visible range rule, which the slice path has to agree with ---
     {
-        const auto gridOf = [](std::vector<float> values) {
+        const auto gridOf = [](std::vector<double> values) {
             amrvis::VolumeGrid grid;
             grid.dims = {static_cast<int>(values.size()), 1, 1};
             grid.region.lower = {{0.0, 0.0, 0.0}};
